feat: Seperate DiffCallback to new file
